move generation code to a Generator module
